New Reaction class for generating Catalysts
The main addition is a description for reactions (such as CO2 + 4 H2 -> Methane + 2 H2O) that allows for the generation of a prototypical Catalyst (That catalyzes that reaction), which can then be mutated and thus produce new Catalysts that have different attributes (but still catalyze the same reaction). These attributes include temperature and pH tolerances, whether the generated Catalyst will be a Ribozyme or Enzyme, what substances inhibit the Catalyst, etc. These MutantCatalysts have to pay for better/ more specialized attributes out of a total pool of points. Thus, the Catalyst has to make tradeoffs for different attributes and therefore every possible mutant naturally has fairly good balance (Each attribute has its own base weight so more important attributes can be further balanced this way). This pool of points is reduced in size if the Catalyst is a Ribozyme, and is inversely proportional to the relative size of its "productivity" attribute, which scales reactants and products, thus controlling space efficiency. The final balancing factor is that the length of the Polymer is directly proportional to the potential (which controls size of pool), so higher potential means more expense in making the Catalyst. To manage the energy balance of reactions, enthalpies on either side of equation are compared, and ATP consumption/ production is automatically generated. Finally, a reactions.json file has been added, that contains all the information necessary to instantiate a Reaction.

To support this addition, a number of changes/ new features were added. Here is an incomplete list:

  - Reaction has been renamed to ReactionAction, to reflect that is is an
    Action, as opposed to the new Reaction class, which actually defines
    the properties of a reaction.
  - ReactionAction now takes Compounds in its constructor, as opposed to
    taking Molecules and just assuming 1 unit of each.
  - Molecules now required to have an Enthalpy (Enthalpy of formation)
  - Compound.Quantity now uses float instead of int
  - Added GenericFunction, which allows functions to be defined that take
    a generic domain value, and output a float. This is used to implement
    classes that randomly sample probablity distributions, such as the normal
    distribution. This is used for mutating the attributes of a Reaction.
  - Added Json library for reading reactions.json.