Solution pH support improvements
pH and pOH now correctly calculated, and addition of protons, hydronium, and hydroxide to a solution now has immediate effect of reacting with water, reacting with hydroxide or reacting with hydronium, respectively. This means that a solution either contains extra (not including autoionization; that's accounted for only when computing pH/pOH) hydroniums OR hydroxides only.

Also adjusted amount of water in organism's cytozol to 1 billion units, to allow for cellular processes to not wildly change pH (may still need to adjust later).

Added bicarbonate and carbonic acid in anticipation of the addition of a generic pH buffering system.