Cell membrane, diffusion, Locales
This update adds support for diffusion across the cell membrane. To support this feature, many changes and additions took place. There is now a class Interface which is responsible for diffusion between two Volumes, another new class. Volume is an interface that requires implementors to define the quantity of any molecule on a square unit of its surface. Solution now implements Volume.

Membrane is a new class that implements Interface between Cytozol and a Locale (described further down) intended to be used by Organism. It contains a permeability dictionary that maps Molecules to their diffision rate modifier.

Now that Organisms need to actually be inside some medium, there was a need for a class to represent that medium, and that class is Locale. Locale represents a extremely small volume, such that its Solution is homogenous. Locales can also contain a small number of Organisms (currently pegged at 25 10-cell organisms being 5% of volume). Locales take the form of different habitats; Water, Land, Air (as well as variations on those, like fresh vs salt water). Locales will in the future be how Organisms come into contact with eachother, and will be nodes in a graph of Locales, each with a matching Interface subclass between them (WaterWaterInterface, WaterLandInterface, etc). Organisms will be able to travel across these boundaries.

With the addition of Locales, it seemed time to reevaluate the role of the World class, which until now was being used simply to contain and run the sole OrganismComponent in the default scene. Its natural role is of course to represent the highest level of organization of the logical objects, and as such, visualization concerns are beyond that scope. So, the visualization was moved to a new class, MicroVisualization. World now simply contains Locales.

Previously, I was unsure of the relationship between the UI and the logical side of the game. Should they be intertwined, to make visualization easy? This is fine for many games, but eventually we would need the ability to quickly compute thousands/millions of Organism lifecycles. The solution to this now seems clear; The logic of game will be wholly independent to the visualization, however, we can visualize logical events of interest as a sort of "replay" that has no effect on the actual logic. And this visualization will be intertwined.

The logical classes that change over time now must implement a new interface, Chronal. This just means they have to implement the function Step(). This function does not supply a time interval, because the idea currently is that each moment of the logic will be a fixed, discrete step. This makes it easier for players to understand the results of their designs, by eliminating uncertainty and making visualization of each step possible.

As a final note, I've set the the number of water molecules in a cell to be 20B, based on a cell volume of 0.6 cubic micrometers. Based on a single cell having 1 unit of surface area and the preceeding volume, I have set up Solution to return 50M for 100% water when calculating QuantityPerArea. This is based on a surface area of 4 square micrometers. Organisms with multiple cells have surface area calculated by # of exposed sides / 6. I've chosen not to include the tops and bottoms of cells, to make the volume-surface area ratio more impactful (organisms are all flat, so without this they wouldn't care about surface area).