Fixing Solution's precision issues
Because Solution deals with up to 100T units of water, we needed high precision math to be able to, for example, add something as low a millionth of a unit. I wanted to avoid simply changing Compound's quantity variable from float to decimal because this was unnecessary for most calculations, and that still doesn't deal with all the conversions that creates in Solution alone. I tried making Solution use decimal internally, but it was very messy because many of the functions it uses internally are public and thus have to take/return floats.

The approach I ended up taking was to create a private internal class called PrecisionSolution, where all math uses decimal, and then I keep an instance of that in each Solution, and expose all the functionality through that instance, essentially making Solution a float wrapper around PrecisionSolution.