Fixed visualization bug
Somehow visualization of MoveCommand became broken 2 commits ago. When MoveCommand was executed after a rotation, the visualization would show the CompoundComponent moving from the pre-rotation position to the pre-rotation destination.

I was not able to determine how the previous commit broke this, but I did determine that the behavior of CellComponent.ValidateSlots() was poorly concieved and that as such the current way animations tracked positions on slots should not have worked at all (again, no idea how it did previously).

The visual slots were, after a logical rotation, being updated such that the visual slot at index 0 would contain the logical slot at index 0. This meant that tracking a visual slot is completely meaningless because it never actually moved (other than the temporary rotation applied to the entire CellComponent).

The better implementation is to have the relationship between the visual slot and the logical slot to be a permanent one, and after a logical rotation, update the local rotation of visual slots to match their new logical index. That's how it works now.