Fixed Catalyst visualization issue
Catalysts were appearing in the DetailPane dna sequences list by the name of the reaction, rather than the name of the catalyst. That has been fixed.

In addition, it was determined during this change that Reaction.Catalyst was creating a new Catalyst with a unique sequence every time it was referenced, so now it only does that the first time, thereafter storing the result.