New Reactions
Added new Reaction data. Included:

 - Hydron sulfide oxidation
 - Glycolysis
 - Nitrogen fixation
 - Imadazole, Purine, Pyrimidine synthesis
 - AMP, CMP, GMP, TMP synthesis

Updated method that processes loads Reactions to use default values in the scenario a value is not present in definition.

Changed "Adenine", "Cytosine", etc in code to "AMP", "CMP", etc to reflect the fact that they are actually nucleotides, not nucleobases. Previous names were kept for the sake of recognizability, and we may still show those names in the UI, but underneath its better to be exact. This was made more important now, because the Reactions data file references AMP, and it really can't call it Adenine because that would be confusing to users seeking to edit the file. Finally, we may one day actually have Adenine, so this prevents future issues.

Text components in dna sequence list were changed to use "best fit" option so that the longer catalyst names render properly.