Reaction brainstorming tool
Added static class ReactionBrainstormer that takes reaction prototypes in the form of a collection of reactants, products, and possible byproducts, and generates all balanced reactions possible, sorted by number of byproducts, then by scale of byproducts.

In order to implement this, I added a number of new Utility and MathUtility functions. This raised the issue of using List by default in these classes, so I switched everything to IEnumerable where applicable.

With this first "tool" (meaning meta functionality; not directly related to the implementation of the game), I decided to future proof and put it inside a new class Tools, which just has a Run() method that goes through and runs all the different tools. For now, the expectation is that each tool simply does some magic (generally starting with reading an input file) and then outputs the results to /Output/.

I ran into more problems with input file errors (previously experienced when I wrote Reaction.LoadFile() and reactions.json), so I added checks to prevent crashes do to this in ReactionBrainstormer.Run(), and then did the same in Reaction.LoadFile() for consistancy. Its a little messy and overly verbose currently, will need to write some powerful json reading utilities at some point.

Added ToolsComponent and attached to object in MicroScene so that the tools can be run by hitting the "r" key.

Discovered that there was an errant "Scene.cs" file in the top level directory of the project (in addition to correct one in Assets), so I removed that.

Discovered that Neon had been listed as "N" rather than "Ne", causing Nitrogen to be interpretted as Neon.

Changed nucleotides to have effective charge of -1 rather than -2 to reflect fact that they only have a acidity of 0.9, which is closer to full dissociation of just one proton.

Finally, updated reactions.json based on above changed to nucleotides and the results of ReactionBrainstormer.