Added Pipase
Added new Ribozyme Pipase for PipeAction and made PipeAction functional. In order to implement PipeAction, I needed a way to get the Slot across from the the pipe Slot (in the adjacent cell), this in turn gave rise to the Slot properties AdjacentCell, AcrossSlot, IsExposed (is a surface edge).

IsExposed was useful for refactoring the implementation of SurfaceArea in Organism, so I updated that. This necessated the ability to enumerate over Slots in a Cell, so I added a new Property Slots to Cell. This however intruded on the responsibility of GetSlot(int index), but replacing it was tricky because that method interpretted indices circularly, so f.e. index 7 meant 1. This makes relative queries much simpler. So, I made a special List subclass for Slots that overrides the index operation. I deleted GetSlot, and updated all references to it.

Finally, I discovered that Slot.AddCompound() actually _sets_ the compound rather than "adding" it. I decided to change the functionality to match the name.