Active transport
Exopumpase and Endopumpase are new Ribozymes that use new class PumpAction to implement active transport across the cell membrane. Whatever compound is placed ontop of these Catalysts will be moved from the cytozol to the surrounding Locale, or vice versa. This process consumes ATP, and its rate is proportional to the concentration of the source.

Separated Action's Beginning() method into Prepare() and Begin(). This allows correct implementation of composite Actions, because all component actions can now be checked before executing any code that mutates the state of game.