Replaced Molecule.CompareMolecule()
Molecule.CompareMolecule() insulted my eyes so I finally replaced it with an .Equals() override. Molecule now compares names first, and in the case where both of the Molecules are not named, compares by elements.

To support the Polymers, which are either unnamed (random DNA segments) or have non-unique names (Ribozymes, Enzymes), Polymer also recieved a .Equals() override. Comparison is simple; must have the same Monomers in the same order.