Added Actuase
Added new Ribozyme Actuase, which moves compounds around based on the DNA strand placed ontop of it. It interprets the DNA as a series of slot location pairs, repeating when it reaches the end of the strand. IT currently moves one unit at a time, and will skip pairs if they fail (rather than wait for success).