Organism splitting
Previous commit implemented detatching of cells, this necessated the ability to detect when a cell's removal causes the organism to exist in multiple pieces, and respond by creating new organisms from the original organism.

Also fixed mistake in last commit where when detatching a cell, it didn't actually remove the cell when creating a new organism.