Try Command
Added new Command TryCommand, whose syntax is [try command codon] [marker] [other command]. TryCommand will attempt to execute the specified command, and if it fails, will instead go to the marker specified. This allows for the player a different way to control the flow of their DNA program. We enforced the now standard policy regarding failed commands, which is to try them again. This was supported (but not rigorously) by Interpretase, but not by Actuase.

To make Interpretase saner, I finally got rid of all the DNA state mutation occuring, and made all the Interpretase methods use index arguments rather than DNA.ActiveCodonIndex except where it made sense.

Part of this rewrite involved separating interpretation from Interpretase.Catalyze() (so that nested commands could be intepretted), and this required a new function that can measure the length of commands. This I used in DetailPane to greatly simplify GetIndendationLevels() and GetCommandGroups().