Support for split catalysts
This commit set out to support the new idea that catalysts should be able to exist in any stack size, but that their productivity would be effected proportionally. This is a fairly simple change if every action could simply be scaled to fit the amount of work that can be done in a single turn, but because there are many actions that are "all or nothing", I needed to come up with a solution that allowed for progress on an action to build between turns.

The solution implemented in this commit is to give actions a cost, which may be scaled (if the action supports it), and to give them a sort of progress bar. Pay(float) increases the progress bar. Catalysts pay for action during Catalyze(), and only when it is fully paid for, do they return it.

Because this functionality is common to all Catalysts and thus needed a common implementation, and because I realized that Catalysts can be Ribozymes or Enzymes arbitrarily, I removed Ribozyme as the parent of all current Catalysts, and replaced it with two abstract classes that implement Catalyst; ProgressiveCatalyst and its child InstantCatalyst. ProgressiveCatalyst implements the behavior described above, making payments on actions until they are paid off, whereas InstantCatalyst pays once and then sets the cost to the payment, thus scaling down the effect of the action.

Now, Ribozyme and Enzyme implement Catalyst by simply containing one of the Catalysts that have functionality of their own.

To store Actions between turns (as they are being paid off), ProgressiveCatalyst now has a static Dictionary that maps Slots to Actions. This is necessary because Molecules are supposed to be singletons (with the exception of DNA).

Two new commands were added, move single (brought back), and move max, which moves an amount equal to stack size of the Catalyst that creates it.

Many other things were changed to fit the new Action cost mechanic (Mainly Actions and Catalysts were affected)

While testing this, I ran into a few bugs that I had to resolve, including an issue where the temperature and ph activity functions were being calculated incorrectly.