Support for moving to across slot, cytozol
Added across slot (slot across from a given slot, in adjacent cell) and cytozol and locations available to player for commands.

Spurred by these new additions, I decided to rethink the current codon groups. Now, A codons are locations (but are interpretted as values contextually), and all T codons are now markers. This improves comprehension because T codons are no longer lumped together with locations, allows players to use 15 different markers rather than only 7, and improves usage of A codons (of which only AAA and AAC were likely to see much use, and even then, rarely). This does cause a visualization issue because now, for example, AAA shows up as "Slot 0", even when used in a value context.

The across slot location is interpretted as the current Locale if the actual across slot is null, and the correct behavior of commands is to add the compound in question to the Locale. This however is not yet supported, instead commands that move to the Locale are simply ignored.

Added DissolveCommand to support moving a compound from a slot to the Cytozol.

Added ActionCommand to allow Commands that reuse Actions.

Added ExposedCompound and RemoveExposedCompound() to Slot, to make it easier/ less error prone to access the topmost Compound of a Slot. I did not enforce this everywhere however, as ignoring the CatalystCompound doesn't really break anything, and because in a very near commit I am completely removing CatalystCompound, so there won't be a topmost Compound. (Situationally having two Compounds on a slot is inelegant and now seeming unecessary. If we do implemented stacked compounds, it will be in the form of bonding compounds together into one).