Removed Slot.CatalystCompound
Prior to this commit, Slots could contain both a Compound and CatalystCompound. This isn't as elegant or simple as just 1 Slot : 1 Compound, but was thought to be necessary to avoid slot space being too constrained. With the addition of the across slot being available to commands, and with catalyst splitting lowering design constraints in general, we are trying out the more obvious system of slots either being empty or containing a single Compound. Supposing this does not work out, we will instead using binding of Compounds/Molecules rather than allowing for two Compounds per slot again.

Because Pipase cannot use its own slot as an input anymore, its functionality has been reimagined as basically a fixed Actuase (the source slot and destination slot are always the same) which is cheaper to make and use. This means its action was simply going to be a generic move action, which doesn't exist, so I decided to pull the functionality out of MoveCommand to a new Action MoveAction, and then made everything use that directly or under the hood. (MoveCommand is now a ActionCommand that uses MoveAction).