Support for moving from Slot to Locale
Prior to this commit, commands to move a compound from a Slot to the Locale would be ignored, as there was no Action nor animation to support this.

In order to implement this, I decided to make a unified class to represent move from a slot to any destination (another Slot, the Cytozol or the Locale), this class replaced MoveAction, which only supported Slot -> Slot.

MoveAction is now a generic class whose template specifies what type the destination is. It is abstract, not implementing End(), instead subclasses MoveToSlotAction, MoveToCytozolAction, and MoveToLocaleAction imlement it. The alternative was to just jam all the functionality into MoveAction and take an object for the destination, but I tend to dislike implementations that revolve around continuously testing an object member. I suppose an interface with AddCompound(Compound) could have also worked, but that is a little messy. In practice, the subclasses stand on their own, as only Interpretase has to determine at runtime which one it wants.

Speaking of which, Interpretase.MoveCommand has been modified to take a Slot source and an object destination, which it then immediately maps into one of the above subclasses (Avoiding the object type hanging around).

Animations for these new Actions was also implemented, with the Locale version's destination position being handled by adding a new GameObject to SlotComponent that resides a little ways outside the SlotComponent.