Volcanic vent WaterLocales, Measures.cs
Added a method to create a WaterLocale with a simple approximation of the composition of of seawater near a volcanic vent. In order to implement this using the data I had in grams per liter, I made a new file, Measures.cs, which allows me to convert between game measurements and real world measurements, as well as to compute derived measurements.

This presented the need to not only to establish exactly what our units meant in relation to real world units, but give them a name. There was already an informal assumption that the in game units for amount of a substance (Compound.Quantity) were not based on mass or volume, but on how many individual molecules you have, so this makes our units a number, like moles. Based on the calculations I did earlier to determine the scale of in game cells (basing them on a model of 0.6 cubic micrometer cell), I chose to make the in game units derived from Moles, and scaled them so that 1 unit in game unit of something would have a relatively small effect on the composition of the cytozol, but that somthing like 100 would have a meaningful impact. In the end, the scale was chose as 10^-18 of a mole, or an attomole. Rather than call our units that, however, as the scale may change in the future (and attomole is too serious in mind to show to the user in a game for a wide audience and one which is not a strict chemical simulation), I decided upon the name "Smole", which is reflects that it is a very small measurement, is lighthearted, and still clearly references Moles, which indicates the basis in chemistry.