Clickable cells
Prior to this commit, there was only one user interface, a window that allowed the user to create a DNA program. When closed, this program was loaded into a fresh organism, the simulation began, and was visualized.

Now, we are looking to take this and move towards implementing one stage in the game, what we might call the "mutation phase", where the player has latitiude to modify the organism. So the target is to have an organism presented immediately, and the player should be able to modify it in many different ways.

To support this, this commit adds the ability for the user to click on cells, and for contextual interfaces to be shown based on this. If the user clicks on a slot with DNA, the DNA viewer/editor discussed above is shown. If the user clicks the center of the cell, a new interface is shown, which displays all the compounds in the cytozol as a tile grid. In the future, these tiles will be drag and droppable, to support futher interaction.

High level overview aside, the actual technical changes are:

 - DetailPane is now DNAPanel

 - A new class, DetailPanel has been added. This class defines a relationship
      between a logical object being displayed and a user interface.
       - DNAPanel and CompoundGrid are its subclasses.
       - As an example, DNAPanel takes a DNA object as its data, and can
         directly edit it.

 - HasDetailPanel is a new interface which requires implementers to supply a
      DetailPanel when asked. The visual representatives of logical objects
      are its implementers.
       - For example, SlotComponent is a HasDetailPane and supplies a DNAPanel
         when clicked while containing DNA.

 - CompoundGrid was added, which displays the contents of its BoundObject

 - CompoundTile was added. This is a UI element that displays information
   about a compound in a square.

 - CellComponent now has a CircleCollider and is clickable. It determines
     which slot was clicked (or whether the cytozol was clicked) and then
     opens the associated DetailPanel

 - New interface IMutableContainer<T> was added. This interface defines a
   container that when changed, can communicate this separately to all
   interested parties. This is necessary so that if mutliple UI elements
   are using a container, they can both determine if they need to update.

 - Solution now implements IMutableContainer<Compound>