Formal Prefabs, Editor-set references
Prior to this commit, I was informally using the prefab pattern by f.e. creating a list with a single element, then instantiating that element to fill the list dyamically. This leads to the need to either turn the first copy into an informal prefab in Awake(), OR in a static function, which then needs to be called every time you might use the prefab. The better alternative was to start using the nice prefab system already provided by unity. This way, the prefab is available immediately, and no lines of code are necessary.

The one downside to the Prefab system in unity is that I had been creating DNAPanels and CompoundGridPanels via static builder methods in those classes, however unity's editor does not allow you to set a static member variable, and thus you are left needing a singleton to place the references into. This would be really awkward for the two classes mentioned, so I finally settled on adding a new script to the scene that contains all the prefabs relevant to the scene, and is accessed like MicroVisualization; through the static class Scene. This script is called MicroPrefabs.

In the spirit of reducing code by using the editor (as well as getting rid of Awake() setup), I've, wherever possible, removed code that connects parts of a UI object (Like DNAPanel and its two child lists codon_list and sequence_list) by using a search, and replaced it by exposing the variable to the editor and then setting the references in the editor.