Card
Card is a new UI element used to visually represent "classes" of in-game objects. The current design is for visualizing Molecules, specifically Catalysts, but Cards are a generic concept that could be extended.

Essentially, every Catalyst can be represented by a Card, and in the future, the player will have a "Deck" of cards, amounting to a their organism's genetic heritage. During mutation, the player will be able to drag cards from their deck to the Organism, to instantiate 1 unit of that Catalyst inside a cell.

Cards have the ability to be inspected more closely to reveal additional information. For example, right now, cards have an image that if clicked zooms in to show a demo of that card's effects. In the future, the full data related to a catalyst will be visible through a similar mechanism.

In order to support this, several new features were added.

  - Example is a new class that contains an Organism and a Length

  - Example is used by new class ExampleComponent, that visualizes
    Example.Length number of steps of Example.Organism's life, then repeats.

  - The Catalyst interface now requires a Description, a Price, and an Example.

  - CatalystPanel, a subclass of DetailPanel was added. This ui element
    contains a Card, and is shown when the player clicks on a catalyst.