Deck
Added a new DetailPanel "DeckPanel". This is for visualizing the genetic resources available to the player, in the form of Catalysts.

Each organism now has a Deck, a collection of Catalysts, which it generates by searching through all its DNA, looking for ribozymes, and enzymes coded for in the DNA. This deck is then used to populate DeckPanel, with each Catalyst represented by a Card (see earlier commit). There is now a deck icon in the corner of the screen which opens the DeckPanel when clicked.