Dragging Cards
Added the ability to drag cards in order to instanciate Compounds.

When you drag a card now, a CompoundTile is created containing a Compound with 1 unit of the Card's Molecule. This CompoundTile can be dropped anywhere for a contextual effect.

Scene.Micro is no longer an internal static class, but a property of scene that fetches a MicroScene instance. This is a new class that I use to store all the references to different singletons in the scene (MicroVisualization, MicroPrefabs, etc). Because its a script in the scene, I can set all these references from the editor. This is preferable to having to do a search to get these references. This change was prompted because I now have 2 Canvases, which means I can no longer rely on simply searching for the only instance of Canvas.

Fixed bug that went on unnoticed for several commits - Even though I got rid of TAA - TCG location codons in the logic and some of the visual layer, I missed many other fragments in the visual layer, resulting in indentation, codon coloring, and codon option errors when using those codons as markers. I believe I've cleared these all up now.

Also fixed another longstanding bug where Polymers cause a crash when calculating mass. This just never occured because I never added a Polymer to a Solution.

Replaced the default StandaloneInputModule with new class MicroInputModule, which contains support for transferring an drag event to another object.