Trashcan
Added Trashcan, TrashcanButton, and TrashcanPanel.

Trashcan serves as a temporary storage location for Compounds while editing an organism, as well as (of course) a way to simply dispose of them. It also serves another, very important purpose; compounds may transform when disposed of in the trashcan, producing new compounds (this currently only works for Catalysts).

Trashcan is represented visually by a button in the corner of the screen (run by TrashcanButton script) which, when clicked, opens TrashcanPanel, a prefab/ script that visualizes the contents of the Trashcan and allows the player to interact with it.

Hole is a new visual element that allows other elements to be put inside it. It masks away the lower half  of the objects to make it look like they are sticking out of a surface. It also controls their movement, smoothly moving elements to the hole, then sliding them slowly into it, as well as preventing them from being moved horizontally while in the hole.