Prefabs for visualization objects
CellComponent, SlotComponent, and CompoundComponent have recieved prefabs that simplify and reduce the amount of code needed.

This also helps modularity, as now that the objects can be seen and modified in the editor, changes can be made and tested much faster and easier.

Slots and the cytozol highlight now have their own colliders, which means I was able to remove the "Part" system from CellComponent.