Undo/Redo
Added functionality to store and restore versions of classes that implement new interface Versionable.

This enables the ability undo/redo actions taken while editing an organism.

Doer is a new class that actually implements this versioning behavior, having the functions Do(), Undo(), and Redo(), with Do() being the instruction to save the current version.

The new script MicroEditor aggregates this functionality so that any number of versionables can be tracked and have their versions manipulated in sync.

MicroVisualization is the first adopter of this functionality. It saves the visualized Organism's version after every step.