Organism versioning bugfix
Organism instances were creating versions of themselves that contained the same reference to any given molecule. This means that changing the properties of one versions molecule changed all versions. This is usually a non issue because molecules are implicitly immutable (if you add a nitrogen to water, its no longer water) so they weren't really part of the organism's state. However, this overlooks DNA, which IS mutable.

To fix this, now Molecule implements Copy(), but the default implementation is simply to return "this". DNA overrides this to return a new DNA molecule (unless the DNA is a Ribozyme, which is not mutable). Organism now uses Molecule.Copy() where applicable in Copy() and Checkout().