Better support for Reaction animations
Reaction animations require larger numbers of compounds moving in/out of slot in the same direction.

This update allows actions animations to spread out the destination/source positions by consuming one of 4 positions on the slot (the three corners and center). It also allows these positions to be shared between separate sections of code, so, for example, the main reaction and the atp reaction don't visually interfere with eachother.