Catalyst state
In preparation for upcoming gameplay changes, Catalysts now have two state variables, a list of "cofactors" (compounds directly attached to catalyst that alter their function), and "orientation", which represents a catalyst's rotation.

The "action in progress" in ProgressiveCatalyst was made unecessary, as we can now store progress, which also means that the parts of Action related to payment were removed.

To support this change, Catalysts had to made Copy()able to play nice with undo/redo system.

Because Ribozymes and Enzymes, being Catalysts, are by extension no londer archetypal, I had to decide between having some sort of system to keep track of the immutable versions of every ribo/enzyme and then instances would be copied from them, or to do away with the whole system. I went with the latter. The only place a central list was being used was in DNAPanel, which was outdated since the introduction of Deck. Its not even used for mutations. Now theres just a dictionary of strings to ribo/enzymes that is used to determine uniqueness both of DNA/Amino acid sequences and the catalyst - length combo.

Ribozyme.GetFamily("Name")[0] instance have been replaced by new Ribozyme(new Name(), x). The arbitrary lengths will be fixed a upcoming commit where Catalysts will have a "complexity" value that will be used to calculate lengths.