Catalyst.Power
The length of a catalyst is now directly proportional to the new property "Power".  This roughly models the reality that more difficult/complex problems require more material/information to solve. Its also a balance tool in that I can offer an incentive to use something by lowering its power, or make something not so good by increasing its power.

Its not just about balance however. For example, Reactions  already had something that amounted to Power, but through mutations the player can actually choose to increase this value because it yields more points available to improve other aspects of the catalyst. The reverse is also true, getting a weaker catalyst with a lower power (and thus shorter length). This tradeoff shows that adjusting power is not just about getting catalysts close the same "power level", but rather about contextualizing; a lightweight organism may want to use a number of low power catalysts, while another may want to make the best use of its available space.