Interpretase changes
Interpretase's command set received a rework. It is now heavily based on self movement and orientation. Movements will cause other compounds to pushed around. Interpretase can now "grab" other compounds in the direction it is facing, causing them to be dragged when it moves. It can now "spin", changing its orientation, and rotating grabbed compounds around itself.

In implementing this, limitations to the current Catalyst and Action designs became apparent and had to be fixed.

Catalyst's now have "facets", which is a way to deal with composition formally through the Catalyst interface. Interpretase has a IsGrabbing state, which needs to be accessed by some of the actions its spawns. These actions use Catalyst.GetFacet<Interpretase>() to do this.

Action.Prepare() was replaced by the IsLegal property. IsLegal is used to filter out actions behind Catalyze() (in addition to the Catalyst simply choosing to return null for whatever reason). Illegal actions are ones that definitely can't execute or can't execute under any arbitrary ordering. Illegality can't be determined at this point for all actions, however, so conflicts can happen and are now supposed to be resolved in Action.End() (Conflict resolution should try implement order agnostic behavior).

The action stages of each step have been formalized and refined. Now the stages are represented by a Action.Stage object, and can be used to determine if a candidate Action fits in a stage. Stages now run completely sequentially, as partial sequentiality would be less comprehensible to users. This means that actions are gathered during every stage (instead of once per step), so that smart sequencing can enable smooth, efficient operation of the organism within a step. To support this, Catalyze() now takes an additional argument; the stage.

Deleted Rotase, Pipase, Actuase, and Sporulase. Rotase is defunct (rotation is now implemented by pushing), and the others all are due to basically be rewritten and I didn't want to bother maintaining them through the changes made to Catalyst and Action made in this commit. Their related ActionComponents were also deleted.

Added Mess, technically a Catalyst Molecule. It's purpose is to reuse Catalyst's cofactor system to implement merged molecules (resulting from a collision between two MoveToSlotActions).