Animations, Attachments, Stackable
The goal of this commit was to add animations for the new actions added recently. In order to do this properly however, it became apparent we needed two new logical features as well; Attachments and Stackable.

Attachments are formalization of the idea that catalysts may have up to 3 orientation based interactions with adjacent slots. Interpretase uses its .Orientation direction (across in rest orientation) for its grabbing functionality. Reaction catalysts may have, for example, two slots it uses as inputs, and one slot it uses as the output. To represent this, now Catalyst has a dictionary of directions to "Attachments", which are subclasses of Attachment. Attachment currently has InputAttachment, OutputAttachment, and Grabber subclasses. Exposing this of course allows for better visualization, and stateful Attachments, such as Grabber, can now expose their state, further supporting visualization.

Before, we determined whether two compounds could be stacked by using Molecule.Equals(). This is actually already problematic because the Undo/Redo system can't detect changes if Equals() returns true for two stackable catalysts with different orientation, for example. But in addition, CompoundComponent couldn't determine if the state of its Compound had changed, requiring updating every single frame, which interferes with some ActionAnimation effects. The solution was to make the Stackable interface, and make .Equals() comprehensive.

With those features out the way, animations for rotating while moving, grabbing, and spinning were added.

To support this, MoleculeComponent (new child object of CompoundComponent), and AttachmentComponent (children of MoleculeComponent) were added.

Finally, Catalyst images were updated to work better with attachments, and to reflect that it is a structural thing rather than just a mass of molecules. It now has a circle around the image representing the molecular structure, and the circle has images of the attachments sticking off of it. Eventually I'd like to add some whimsical attachments like a hammer for cell construction, a saw for cell separation, etc.